P4-2449 Move PER logic to assessmentable concern
jira-ticket: https://dsdmoj.atlassian.net/browse/P4-2449

* For reuse across multiple assessments, move PER model logic to a
concern, and move tests to a shared example.

* Rename state machine to a generic framework assessment name
* Ignore rubocop warning as its a false positive
