+++ spec/factories/person_escort_record.rb
@@ -21,16 +21,16 @@ FactoryBot.define do
     end
 
     trait :in_progress do
-      status { PersonEscortRecord::ASSESSMENT_IN_PROGRESS }
+      status { FrameworkAssessmentable::ASSESSMENT_IN_PROGRESS }
     end
 
     trait :completed do
-      status { PersonEscortRecord::ASSESSMENT_COMPLETED }
+      status { FrameworkAssessmentable::ASSESSMENT_COMPLETED }
       completed_at { Time.zone.now }
     end
 
     trait :confirmed do
-      status { PersonEscortRecord::ASSESSMENT_CONFIRMED }
+      status { FrameworkAssessmentable::ASSESSMENT_CONFIRMED }
       confirmed_at { Time.zone.now }
     end
